Shinytest2 improvements and unit testing setup better

+++ tests/testthat/setup.R
@@ -2,36 +2,87 @@
 # SETUP FOR UNIT TESTS ####
 
 ############################### #
-cat("\n\n\n               !!!!!!!!!!!!!! Starting setup.R for testing !!!!!!!!!!!!!! \n\n\n")
+if (!isTRUE(getOption("EJAM.test_setup_banner_shown"))) {
+  options(EJAM.test_setup_banner_shown = TRUE)
+  cat("\n\n\n               !!!!!!!!!!!!!! Starting setup.R for testing !!!!!!!!!!!!!! \n\n\n")
+}
 
 # # This script SHOULD get run before tests, so fixtures created here will be available to all the tests.
 # The file now DOES do library(EJAM) - that would do .onAttach() and dataload_dynamic() and indexblocks()
+# ############################### #
+# cat("\n\n      ------------------ NOW DOING library(EJAM) !!!!  -------------      \n\n")
+#
+if (!"package:EJAM" %in% search()) {
+  library(EJAM)
+}
 ############################### #
-cat("\n\n      ------------------ NOW DOING library(EJAM) !!!!  -------------      \n\n")
 
-# library(EJAM) ####
-library(EJAM)
-############################### #
 
-# >> not load_all() ??? ####
+# do not address install_local() or load_all() in setup.R since setup.R also is used for check() /  R CMD check
+# and you assume it is assuming installed package in that case.
+#
+# Instead handle whether to do install_local() or at least load_all() when interactively starting tests, like via test_ejam()
 
-message("Consider whether you want to use remotes::install_local() or devtools::load_all() before running tests...
- It is recommended during development to use `remotes::install_local()`
- to ensure your development code is the one tested, especially for the webapp functionality tests.
- This is because shinytest2 automatically references the installed version of a package.
 
- Also, if you just use require() or library() here and do not use load_all(),
- then tests will not have access to internal functions like latlon_infer()
- so those tests (if run interactively at least) fail,
- and even if a test were changed to say EJAM:::latlon_infer(),
- that would still only test the installed version, not the source version, which may differ.
-")
+# >>  load_all() ??? ####
+
+# message("Consider whether you want to use remotes::install_local() or devtools::load_all() before running the set of tests...
+#  It is recommended during development to use `remotes::install_local()` or `devtools::load_all()`
+#  to ensure your development code is the one tested.
+#  This is because test_file(), which is used by test_ejam(), uses the installed version of a package.
+#  Likewise, shinytest2 automatically references the installed version of a package.
+#  If you tried to use something like devtools::test(filter='^ejamapi$') for each test file,
+#  that would use the source version of the package, but be very slow since it
+#  would repeat load_all() for every file it tested.
+#  It would be ok to do devtools::test() to test all files, though, if
+#  testing all the test files in the local source version, since it would do
+#  load_all() just once and then run all the tests.
+#
+#  Also, if you just use require() or library() here and do not use load_all(),
+#  then tests will not have access to internal functions like latlon_infer()
+#  so those tests (if run interactively at least) fail,
+#  and even if a test were changed to say EJAM:::latlon_infer(),
+#  that would still only test the installed version, not the source version, which may differ.
+# ")
+
+#
+# # Do you want to install the package now from latest local source? (shinytest2 or always tests on the installed version)
+#
+# if (interactive()) {
+#   install_now = FALSE
+#   # cat("THIS WILL TEST THE LAST-INSTALLED VERSION - IF YOU WANT TO TEST THE LOCAL SOURCE VERSION, DO
+#   #     remotes::install_local('.', force = T, upgrade = 'never', build = F, build_vignettes = F, build_manual = F, dependencies = F)
+#   #     FIRST ! \n")
+# } else {
+#   install_now = FALSE
+# }
+#
+# if (install_now) {
+#   if (file.exists("DESCRIPTION")) {
+#     remotes::install_local('.', force = T, upgrade = "never", build = F, build_vignettes = F, build_manual = F, dependencies = F)
+#   } else {
+#     if (file.exists("../DESCRIPTION")) {
+#       remotes::install_local('..', force = T, upgrade = "never", build = F, build_vignettes = F, build_manual = F, dependencies = F)
+#     } else {
+#       stop("cannot do remotes::install_local() since cannot find source directory")
+#     }
+#   }
+# }
 ############################### #
 
 # require key packages ####
 
 # To run tests interactively, and maybe to test webapp functionality,
 # you need to load() or require() various packages!
+
+# If you start from clean R session and  just did load_all() only, then it would not do all these,
+# and would not do the .onAttach() global_defaults* etc.??
+
+# Maybe the safest in test_ejam() is to do install_local(), library(EJAM) just to be sure, and then ALSO do load_all() so ::: not needed.
+
+# # If you installed EJAM   that should have already checked all these
+# If you did library(EJAM) that should have access to these pkgs but would NOT actually do library() or require() on   all these.
+# If you did install_local() same.
 
 if (!require(testthat))   {cat("Need testthat package for unit tests to work \n\n")}
 if (!require(mapview))    {cat("Need mapview package for some tests of mapping to work \n\n")}
@@ -41,18 +92,10 @@ if (!require(rmarkdown))        {cat("Need rmarkdown  package for some tests to 
 if (!require(data.table))       {cat("Need data.table package for some tests to work \n\n")}
 if (!require(magrittr))         {cat("Need magrittr   package for some tests to work \n\n")}
 ############################### #
-## no simple way to check if doing tests of webapp, or just other unit tests, so just load these here in case needed, even though it slows it down:
-  # if doing tests of webapp, need shinytest2 package loaded
-  if (!require(shinytest2)) {cat("Need shinytest2 package for some tests of web app to work \n\n")}
-  # if doing tests of webapp, need the function from this file, so source it here
-  testdir = testthat::test_path()
-  if (!exists("shinytest2_webapp_functionality")) {
-    if (file.exists(file.path(testdir, "setup-shinytest2.R"))) {
-      source(file.path(testdir, "setup-shinytest2.R"))
-    } else {
-      message("Need to source the setup-shinytest2.R file first to test webapp functionality \n")
-    }
-  }
+
+# anything that runs tests, such as testthat::test_file() done by test_ejam_bygroup()
+
+
 ############################### #
 # is internet available? ####
 EJAM:::offline_warning("NO INTERNET CONNECTION AVAILABLE - SOME TESTS MAY FAIL WITHOUT CLEAR EXPLANATION")
@@ -64,10 +107,14 @@ EJAM:::offline_cat("\n\nNO INTERNET CONNECTION AVAILABLE - SOME TESTS MAY FAIL W
 #  changed global options, a side effect we probably want functions to avoid.
 # but Warm up readr before state inspection starts. The first readr::read_csv()
 # call creates session options like readr.default_locale/readr.num_threads.
-tf <- tempfile(fileext = ".csv")
-writeLines(c("x", "1"), tf)
-invisible(readr::read_csv(tf, show_col_types = FALSE))
-unlink(tf)
+if (!isTRUE(getOption("EJAM.readr_warmed_for_tests"))) {
+  options(EJAM.readr_warmed_for_tests = TRUE)
+
+  tf <- tempfile(fileext = ".csv")
+  writeLines(c("x", "1"), tf)
+  invisible(readr::read_csv(tf, show_col_types = FALSE))
+  unlink(tf)
+}
 
 testthat::set_state_inspector(function() {
   list(options = options())
@@ -77,16 +124,20 @@ testthat::set_state_inspector(function() {
 # (some are normally only loaded into memory if / when needed)
 ## (and build index, but library(EJAM) should have already done the indexblocks() part at least)
 
-# suppressMessages({suppressWarnings({
-dataload_dynamic("all",  # needs frs, etc.
-                 # folder_local_source = file.path(.libPaths()[1],'EJAM','data'), # if installed by testthat in special folder then those are not available
-                 silent = FALSE
-)
-# })})
-if (!exists("frs")) {stop('needs frs etc.')}
-suppressMessages({suppressWarnings({
-  indexblocks()
-})})
+if (!isTRUE(getOption("EJAM.test_data_setup_done"))) {
+  options(EJAM.test_data_setup_done = TRUE)
+
+  # suppressMessages({suppressWarnings({
+  dataload_dynamic("all",  # needs frs, etc.
+                   # folder_local_source = file.path(.libPaths()[1],'EJAM','data'), # if installed by testthat in special folder then those are not available
+                   silent = FALSE
+  )
+  # })})
+  if (!exists("frs")) {stop('needs frs etc.')}
+  suppressMessages({suppressWarnings({
+    indexblocks()
+  })})
+}
 ############################### #
 # global defaults   ####
 
@@ -94,6 +145,41 @@ suppressMessages({suppressWarnings({
 ## When tests try to test the shiny app, the ejamapp() function would normally handle using global_defaults_*.R
 # (The  global defaults also were  being directly put in envt by  /tests/testthat.R BUT THAT FILE MIGHT NOT BE USED ANY MORE )
 #
+################################# # ################################# #
+
+# THIS WAS HOW testthat.R was grabbing the  global defaults but probably not needed here ####
+
+################################# # ################################# #
+
+## this also needs to load EXTRA global_defaults_ info
+# some of which only get checked after ejamapp() done. Not sure that works the way shinytest2 testing is done here.
+# ## that is normally saved in golem_opts during app_run() like...
+#
+# global_defaults_or_user_options <- EJAM:::get_global_defaults_or_user_options(
+#   user_specified_options = list(), # list(...),
+#   bookmarking_allowed = "disable" # enableBookmarking
+# )
+# # app$appOptions$golem_options <- global_defaults_or_user_options
+# ## but shinytest2 does not use app_run() so we need to do it here somehow...
+# ## Try  saving the objects in the global envt ....
+# ## but global_or_param() would find them but get_golem_options() would not?
+# ## some had been done via get_golem_options() like these:
+# ## could switch to use global_or_param() for those places in ui and server.
+# ##  found  golem::get_golem_options("   replaced with EJAM:::global_or_param("
+# # use_shapefile_from_any
+# # default_shp_oktypes_1
+# # default_extratable_list_of_sections
+# # ejam_app_version
+# # default_extratable_hide_missing_rows_for
+# ## assign each global default value to this envt :
+# # see also EJAM:::get_global_defaults_or_user_options()
+#
+# for (i in seq_along(global_defaults_or_user_options)) {
+#   assign(names(global_defaults_or_user_options)[i], (global_defaults_or_user_options[[i]]))
+# }
+
+
+
 ## but note app.R would normally use isPublic and hide the Advanced tab and some options like MACT/SIC/EPA program
 ## so isPublic can be set FALSE here and that should be found by app.R when shinytest2 tests start and use app.R
